chore: Prepare for Tailwind 2.0 - Part 3 (#80)
* drop comment

* run tests through the whole tailwindcss process

This allows us to run against the real implementation. This allows us to actually look at the final output. Added benefit is that we don't have to mock certain plugin api implementations

* simplify customForms implementation

* add a css layer, to be consistent with other plugins

This also allows us to add other keys in the future if we need them without introducing a breaking change

* bump tailwindcss version

* use lodash methods for older node compatibility